updating user object.
